requirements.txt
numpy==1.26.4
POT==0.9.5
scikit-learn==1.3.2
scipy==1.11.4
active-semi-supervised-clustering==0.0.1
matplotlib==3.9.4
PyYAML==6.0.2
pygam==0.9.1
torch==2.6.0
umap-learn==0.5.7
pamona==0.1.0
tqdm==4.67.1